Finished refactoring offers prop-types
