+++ src/components/place-card/place-card.e2e.test.js
@@ -6,7 +6,7 @@ import PlaceCard from './place-card.jsx';
 Enzyme.configure({adapter: new Adapter()});
 
 it(`mouseover and mouseleave for PlaceCard`, () => {
-  const offer = {image: `image`, priceValue: 100, priceText: `text`, name: `name`, type: `type`};
+  const offer = {image: `image`, priceValue: 100, priceText: `text`, name: `name`, type: `type`, coordinates: [1, 2]};
   const onMouseEnter = jest.fn();
   const onMouseLeave = jest.fn();
   const wrapper = shallow(<PlaceCard offer={offer} onMouseEnter={onMouseEnter} onMouseLeave={onMouseLeave}/>);
