+++ src/components/place-card/place-card.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import renderer from 'react-test-renderer';
 import PlaceCard from './place-card.jsx';
 
-const offer = {image: `image`, priceValue: 100, priceText: `text`, name: `name`, type: `type`};
+const offer = {image: `image`, priceValue: 100, priceText: `text`, name: `name`, type: `type`, coordinates: [1, 2]};
 
 it(`renders correct PlaceCard component`, () => {
   const onMouseEnter = jest.fn();
