+++ src/components/place-cards/place-cards.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import renderer from 'react-test-renderer';
 import PlaceCards from './place-cards.jsx';
 
-const offers = [{image: `image`, priceValue: 100, priceText: `text`, name: `name`, type: `type`}];
+const offers = [{image: `image`, priceValue: 100, priceText: `text`, name: `name`, type: `type`, coordinates: [1, 2]}];
 
 it(`renders correct PlaceCards component`, () => {
   const tree = renderer.create(<PlaceCards offers={offers} />).toJSON();
